+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [FIX] Propagate native `anonymous_id` to WebView RUM and Log events. See [#2847][]
 - [IMPROVEMENT] Add device properties `isLowRam`, `logicalCpuCount`, `totalRam` to the `LogEvent` Objective-C API. See [#2854][]
+- [IMPROVEMENT] Add `collectAccessibility` property to the Objective-C API of RUM configuration. See [#2855][]
 
 # 3.10.0 / 16-04-2026
 
@@ -1129,6 +1130,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2829]: https://github.com/DataDog/dd-sdk-ios/pull/2829
 [#2847]: https://github.com/DataDog/dd-sdk-ios/pull/2847
 [#2854]: https://github.com/DataDog/dd-sdk-ios/pull/2854
+[#2855]: https://github.com/DataDog/dd-sdk-ios/pull/2855
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin