+++ CHANGELOG.md
@@ -1130,6 +1130,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2829]: https://github.com/DataDog/dd-sdk-ios/pull/2829
 [#2847]: https://github.com/DataDog/dd-sdk-ios/pull/2847
 [#2854]: https://github.com/DataDog/dd-sdk-ios/pull/2854
+[#2845]: https://github.com/DataDog/dd-sdk-ios/pull/2845
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin